Fixed being able to add a shoe to a the list.
Able to now add a shoe to the list which is publically exposed for observing changes in the ShoeLIstingFragment

If a shoe is added then it is navigated back to the ShoeListingFragment.

Next step is to render each shoe in the UI.

This the last piece of major functionality in the app before wrap up tasks.