POC Binding
Build a small POC binding with some hard coded values as there is a type issue with the shoeBinding to loop through the list.

I just wanted to ensure I could assign some form of a value and have it render inside of the textviews before styling them or fixing the type incompatiblity.

I am going to fix the type incompatibility next.